+++ apps/sponsor/admin.py
@@ -4,12 +4,12 @@ from django.conf import settings
 from django.contrib import admin
 from django.utils.translation import ugettext_lazy as _
 
-from apps.sponsor.models import Sponsor, SponsorRelation, SponsorTranslation
+from apps.sponsor.models import Sponsor, SponsorRelation, SponsorRelationTranslation, SponsorTranslation
 
 
 class SponsorTranslationInlineAdmin(admin.StackedInline):
-    verbose_name = _(u'Translation')
-    verbose_name_plural = _(u'Translations')
+    verbose_name = _(u'partner translation')
+    verbose_name_plural = _(u'partner translations')
     model = SponsorTranslation
     max_num = len(settings.LANGUAGES)
     extra = 2
@@ -21,11 +21,20 @@ class SponsorAdmin(admin.ModelAdmin):
     inlines = [SponsorTranslationInlineAdmin]
 
 
+class SponsorRelationTranslationInlineAdmin(admin.StackedInline):
+    verbose_name = _(u'partner relation translation')
+    verbose_name_plural = _(u'partner relation translations')
+    model = SponsorRelationTranslation
+    max_num = len(settings.LANGUAGES)
+    extra = 2
+
+
 class SponsorRelationAdmin(admin.ModelAdmin):
     model = SponsorRelation
     list_display = ('__unicode__', 'priority')
     list_filter = ('lan',)
     search_fields = ['sponsor__title']
+    inlines = [SponsorRelationTranslationInlineAdmin]
 
 
 admin.site.register(Sponsor, SponsorAdmin)