+++ CHANGELOG.md
@@ -17,6 +17,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added separate option for LAN media on front page.
 - Added user info on ticket purchase page, to force users to check their info.
 - Added LAN option for allowing manual payments on arrivals page.
+- Added partner relation descriptions, in addition to the existing partner descriptions.
 
 ### Changed